requirements.txt
Pillow>=10.0.0
streamlit>=1.32.0
